requirements.txt
blinker==1.9.0
click==8.3.1
dlib @ file:///Users/runner/miniforge3/conda-bld/dlib_1697537213499/work
face-recognition==1.3.0
face_recognition_models==0.3.0
Flask==3.1.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy @ file:///Users/runner/miniforge3/conda-bld/numpy_1707225416355/work/dist/numpy-1.26.4-cp310-cp310-macosx_11_0_arm64.whl#sha256=14c628943f67e9c10d22faa9c1cdcbf69c3192e88f2f597d03b77a252e589567
opencv-python==4.13.0.92
packaging @ file:///opt/miniconda3/conda-bld/packaging_1761049079158/work
pillow==12.1.1
Werkzeug==3.1.6

numpy